+++ lib/unleashed/models/company.rb
@@ -1,0 +1,5 @@
+module Unleashed
+  # Manage Companies
+  class Company < BaseModel
+  end
+end